+++ initialize.py
@@ -22,6 +22,19 @@ def download_bert_models():
                 )
 
 
+def download_slm_model():
+    local_path = Path("slm/wavlm-base-plus/")
+    file = "pytorch_model.bin"
+    if not Path(local_path).joinpath(file).exists():
+        logger.info(f"Downloading wavlm-base-plus {file}")
+        hf_hub_download(
+            "microsoft/wavlm-base-plus",
+            file,
+            local_dir=local_path,
+            local_dir_use_symlinks=False,
+        )
+
+
 def download_pretrained_models():
     files = ["G_0.safetensors", "D_0.safetensors", "DUR_0.safetensors"]
     local_path = Path("pretrained")
@@ -30,6 +43,20 @@ def download_pretrained_models():
             logger.info(f"Downloading pretrained {file}")
             hf_hub_download(
                 "litagin/Style-Bert-VITS2-1.0-base",
+                file,
+                local_dir=local_path,
+                local_dir_use_symlinks=False,
+            )
+
+
+def download_jp_extra_pretrained_models():
+    files = ["G_0.safetensors", "D_0.safetensors", "WD_0.safetensors"]
+    local_path = Path("pretrained_jp_extra")
+    for file in files:
+        if not Path(local_path).joinpath(file).exists():
+            logger.info(f"Downloading JP-Extra pretrained {file}")
+            hf_hub_download(
+                "litagin/Style-Bert-VITS2-2.0-base-JP-Extra",
                 file,
                 local_dir=local_path,
                 local_dir_use_symlinks=False,
@@ -64,6 +91,10 @@ def download_jvnv_models():
 
 download_bert_models()
 
+download_slm_model()
+
 download_pretrained_models()
+
+download_jp_extra_pretrained_models()
 
 download_jvnv_models()